Move log initialization to separate file
